lama/requirements.txt
# https://pypi.org/project/
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# set PYTHONPATH=%PYTHONPATH%;.
# python bin/predict.py model.path=E:\huangyifu\lama\big-lama indir=E:\huangyifu\lama\test outdir=E:\huangyifu\lama\outputs  refine=True device=gpu
pyyaml
tqdm
numpy==1.24
easydict==1.9.0
scikit-image>=0.20.0
scikit-learn>=1.1.3
opencv-python
#tensorflow
joblib
matplotlib
pandas
albumentations==0.5.2
hydra-core==1.3.2
pytorch-lightning==1.2.9
tabulate
kornia==0.5.0
webdataset
packaging
wldhx.yadisk-direct
torch>=2.0.0
torchvision>=0.15.1
